fix(onboarding): reuse library setup flow
